requirements.txt
streamlit
plotly
pandas
pillow
openpyxl 